requirements.txt
accelerate=1.0.1=pypi_0
addict=2.4.0=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.10=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.5.2=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
beautifulsoup4=4.12.3=pypi_0
bitsandbytes=0.44.1=pypi_0
ca-certificates=2024.9.24=h06a4308_0
cbor=1.0.0=pypi_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.0=pypi_0
clang=17.0.6=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contexttimer=0.3.3=pypi_0
contourpy=1.1.1=pypi_0
cryptography=43.0.3=pypi_0
cycler=0.12.1=pypi_0
datasets=2.19.0=pypi_0
debugpy=1.6.7=py38h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
deprecated=1.2.14=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
einops=0.8.0=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.1.0=pyhd8ed1ab_0
filelock=3.13.1=pypi_0
fire=0.7.0=pypi_0
flagembedding=1.3.3=pypi_0
flash-attn=2.7.0.post2=pypi_0
fonttools=4.54.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.6=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.26.2=pypi_0
idna=3.10=pypi_0
ijson=3.3.0=pypi_0
importlib-metadata=8.5.0=pyha770c72_0
importlib-resources=6.4.5=pypi_0
iniconfig=2.0.0=pypi_0
inscriptis=2.5.0=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.12.2=pyh41d4057_0
ipywidgets=8.1.5=pypi_0
ir-datasets=0.5.9=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
jiter=0.6.1=pypi_0
joblib=1.4.2=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.7=pypi_0
krb5=1.21.3=h143b758_0
ld_impl_linux-64=2.40=h12ee557_0
libclang=18.1.1=pypi_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgomp=14.2.0=h77fa898_1
libsodium=1.0.20=h4ab18f5_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=11.2.0=h1234567_1
lxml=5.3.0=pypi_0
lz4=4.3.3=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
modelscope=1.20.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.0=pypi_0
numpy=1.24.1=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.1.105=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openai=1.52.2=pypi_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.0.3=pypi_0
parso=0.8.4=pyhd8ed1ab_0
peft=0.13.2=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=24.2=py38h06a4308_0
platformdirs=4.3.6=pyhd8ed1ab_0
pluggy=1.5.0=pypi_0
ply=3.11=pypi_0
prompt-toolkit=3.0.48=pyha770c72_0
prompt_toolkit=3.0.48=hd8ed1ab_0
propcache=0.2.0=pypi_0
protobuf=5.28.3=pypi_0
psutil=6.0.0=py38hfb59056_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyarrow=17.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pycparser=2.22=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pygithub=2.4.0=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyjwt=2.9.0=pypi_0
pynacl=1.5.0=pypi_0
pyparsing=3.1.4=pypi_0
pytest=8.3.3=pypi_0
python=3.8.20=he870216_0
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.8=2_cp38
pytz=2024.2=pypi_0
pyverilog=1.3.0=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=25.1.2=py38h6a678d5_0
readline=8.2=h5eee18b_0
regex=2024.9.11=pypi_0
requests=2.32.3=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
sentence-transformers=3.2.1=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=75.1.0=py38h06a4308_0
simplejson=3.19.3=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.1=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.7.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
tomli=2.1.0=pypi_0
torch=2.4.1+cu121=pypi_0
torch-geometric=2.6.1=pypi_0
torchaudio=2.4.1+cu121=pypi_0
torchvision=0.19.1+cu121=pypi_0
tornado=6.4.1=py38hfb59056_0
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.44.2=pypi_0
trec-car-tools=2.6=pypi_0
triton=3.0.0=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.2=pypi_0
unlzw3=0.2.2=pypi_0
urllib3=2.2.3=pypi_0
verilog-eval=1.0=dev_0
warc3-wet=0.2.5=pypi_0
warc3-wet-clueweb09=0.2.5=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.44.0=py38h06a4308_0
widgetsnbextension=4.0.13=pypi_0
wrapt=1.16.0=pypi_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.15.2=pypi_0
zeromq=4.3.5=h3b0a872_6
zipp=3.20.2=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zlib-state=0.1.9=pypi_0
